+++ main/manager.c
@@ -2455,7 +2455,7 @@ static int authenticate(struct mansession *s, const struct message *m)
 			MD5Update(&md5, (unsigned char *) user->secret, strlen(user->secret));
 			MD5Final(digest, &md5);
 			for (x = 0; x < 16; x++)
-				len += sprintf(md5key + len, "%2.2x", digest[x]);
+				len += sprintf(md5key + len, "%2.2x", (unsigned)digest[x]);
 			if (!strcmp(md5key, key)) {
 				error = 0;
 			} else {
@@ -3378,7 +3378,7 @@ static int action_status(struct mansession *s, const struct message *m)
 			"ConnectedLineNum: %s\r\n"
 			"ConnectedLineName: %s\r\n"
 			"Accountcode: %s\r\n"
-			"ChannelState: %d\r\n"
+			"ChannelState: %u\r\n"
 			"ChannelStateDesc: %s\r\n"
 			"Context: %s\r\n"
 			"Extension: %s\r\n"
@@ -4610,7 +4610,7 @@ static int action_coreshowchannels(struct mansession *s, const struct message *m
 			"Context: %s\r\n"
 			"Extension: %s\r\n"
 			"Priority: %d\r\n"
-			"ChannelState: %d\r\n"
+			"ChannelState: %u\r\n"
 			"ChannelStateDesc: %s\r\n"
 			"Application: %s\r\n"
 			"ApplicationData: %s\r\n"
@@ -5368,7 +5368,7 @@ int ast_manager_unregister(char *action)
 		while (cur->active_count) {
 			if (5 <= time(NULL) - now) {
 				ast_debug(1,
-					"Unregister manager action %s timed out waiting for %d active instances to complete\n",
+					"Unregister manager action %s timed out waiting for %u active instances to complete\n",
 					action, cur->active_count);
 				break;
 			}
